pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cloudscraper>=1.2.71",
    "faiss-cpu>=1.13.2",
    "google-genai>=1.68.0",
    "gtts>=2.5.4",
    "lxml>=6.0.2",
    "numpy>=2.4.3",
    "openai-whisper>=20250625",
    "playwright>=1.58.0",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "sounddevice>=0.5.5",
    "torch>=2.11.0",
]
